pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "interviewer"
version = "0.1.0"
description = "Standalone AI interviewer with symmetric LLM agents"
requires-python = ">=3.10"
dependencies = ["openai>=1.0.0", "pydantic>=2.0.0", "python-dotenv>=1.0.0"]

[project.optional-dependencies]
cli = ["typer>=0.9.0", "rich>=13.0.0"]
server = ["fastapi>=0.100.0", "uvicorn[standard]>=0.20.0", "python-dotenv>=1.0.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23", "httpx>=0.25", "ruff>=0.1"]

[project.scripts]
interview = "cli.main:app"

[tool.setuptools.packages.find]
include = ["interviewer*", "cli*", "server*"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
